feat: add new pricing structure
