+++ src/pages/pricing/Head/index.tsx
@@ -1,0 +1,16 @@
+import React from "react"
+
+import Seo from "../../../components/seo"
+
+export const Head = ({ data: { post } }) => {
+    return (
+        <Seo
+            title={"Pricing"}
+            description={
+                "Reduce your data costs and latency with managed streaming CDC and ETL pipelines."
+            }
+        />
+    )
+}
+
+export default Head
