+++ src/style.less
@@ -1777,11 +1777,6 @@ table th {
 .pricing-page-background-image-wrapper {
     display: flex;
     flex-direction: column;
-    background-size: contain;
-    width: 100%;
-    background-image: url("./images/product-flow-section-one-background.svg");
-    background-repeat: no-repeat;
-    background-size: contain;
 }
 
 .pricing-landing-image {
